#42 Chore: 뉴스검색 페이지 경로 /search 로 변경
